contract-chat-filler/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
pydantic==2.9.1
pdfminer.six==20231228
PyPDF2==3.0.1
lxml==5.3.0
rapidfuzz==3.9.7
